+++ selfdrive/debug/count_events.py
@@ -7,7 +7,7 @@ from pprint import pprint
 from tqdm import tqdm
 from typing import List, Tuple, cast
 
-from cereal.services import service_list
+from cereal.services import SERVICE_LIST
 from openpilot.tools.lib.route import Route
 from openpilot.tools.lib.logreader import LogReader
 
@@ -17,7 +17,7 @@ if __name__ == "__main__":
   cnt_valid: Counter = Counter()
   cnt_events: Counter = Counter()
 
-  cams = [s for s in service_list if s.endswith('CameraState')]
+  cams = [s for s in SERVICE_LIST if s.endswith('CameraState')]
   cnt_cameras = dict.fromkeys(cams, 0)
 
   alerts: List[Tuple[float, str]] = []
@@ -62,7 +62,7 @@ if __name__ == "__main__":
   print("\n")
   print("Cameras")
   for k, v in cnt_cameras.items():
-    s = service_list[k]
+    s = SERVICE_LIST[k]
     expected_frames = int(s.frequency * duration / cast(float, s.decimation))
     print("  ", k.ljust(20), f"{v}, {v/expected_frames:.1%} of expected")