+++ tools/cabana/settings.h
@@ -1,11 +1,13 @@
 #pragma once
 
+#include <QApplication>
 #include <QByteArray>
 #include <QCheckBox>
 #include <QComboBox>
 #include <QDialog>
 #include <QGroupBox>
 #include <QLineEdit>
+#include <QSettings>
 #include <QSpinBox>
 
 #define LIGHT_THEME 1
@@ -22,9 +24,10 @@ public:
     AlwaysBE,
   };
 
-  Settings();
-  void save();
+  Settings() {}
+  QSettings::Status save();
   void load();
+  inline static QString filePath() { return QApplication::applicationDirPath() + "/settings"; }
 
   int fps = 10;
   int max_cached_minutes = 30;